+++ src/hooks/usePreviousValue.ts
@@ -1,13 +1,11 @@
 import { useEffect, useRef } from 'react';
 
-const usePreviousValue = value => {
-  const previousValueRef = useRef();
+export default function usePreviousValue<T>(value: T) {
+  const previousValueRef = useRef<T>();
 
   useEffect(() => {
     previousValueRef.current = value;
   }, [value]);
 
   return previousValueRef.current;
-};
-
-export default usePreviousValue;
+}
